fix: support t_Co greater than 256
&termguicolors is computed asynchronously in Neovim, so it can happen that the
check on &termguicolors is falsy by the time the colorscheme is loaded. When
this is the case, the t_Co value is the only way we have left to determine
whether the colorscheme should be loaded on not.

While most modern terminfos expose 'colors#256' by convention and use the
'RGB' flag to announce their support for 24-bit RGB color ("True colors"), the
*-direct variants of those terminfos tend to expose 'colors#0x1000000'
(16777216) instead[1]. We need to treat this as a valid value.

Note that 24-bit RGB color gets enabled in the TUI as soon as &termguicolors
is computed, so there is no need to take the timing into consideration while
loading the colorscheme; colors are displayed correctly as soon terminal
features are detected, even if degraded to 256 colors at first.

[1]: https://stackoverflow.com/a/63611791